Move rules code to the rules module
