+++ app/src/test/kotlin/org/virtuslab/bazelsteward/app/HeuristicTest.kt
@@ -8,9 +8,9 @@ import org.virtuslab.bazelsteward.core.common.FileChange
 import org.virtuslab.bazelsteward.core.common.TextFile
 import org.virtuslab.bazelsteward.core.common.UpdateSuggestion
 import org.virtuslab.bazelsteward.core.library.SemanticVersion
-import org.virtuslab.bazelsteward.core.replacement.Heuristic
 import org.virtuslab.bazelsteward.core.replacement.LibraryUpdateResolver
 import org.virtuslab.bazelsteward.core.replacement.VersionOnlyHeuristic
+import org.virtuslab.bazelsteward.core.replacement.VersionReplacementHeuristic
 import org.virtuslab.bazelsteward.core.replacement.WholeLibraryHeuristic
 import org.virtuslab.bazelsteward.maven.MavenCoordinates
 import org.virtuslab.bazelsteward.maven.MavenLibraryId
@@ -240,7 +240,7 @@ class HeuristicTest {
   private fun resolveUpdates(
     library: MavenCoordinates,
     version: SemanticVersion,
-    vararg heuristics: Heuristic = allHeuristics
+    vararg heuristics: VersionReplacementHeuristic = allHeuristics
   ): FileChange? {
     return resolver.resolve(files, UpdateSuggestion(library, version), heuristics.toList())?.fileChanges?.firstOrNull()
   }
