+++ core/src/main/kotlin/org/virtuslab/bazelsteward/core/DependencyKind.kt
@@ -2,12 +2,12 @@ package org.virtuslab.bazelsteward.core
 
 import org.virtuslab.bazelsteward.core.library.Library
 import org.virtuslab.bazelsteward.core.library.Version
-import org.virtuslab.bazelsteward.core.replacement.Heuristic
+import org.virtuslab.bazelsteward.core.replacement.VersionReplacementHeuristic
 import java.nio.file.Path
 
 interface DependencyKind<Lib : Library> {
   val name: String
   suspend fun findAvailableVersions(workspaceRoot: Path): Map<Lib, List<Version>>
   val defaultSearchPatterns: List<PathPattern>
-  val defaultVersionDetectionHeuristics: List<Heuristic>
+  val defaultVersionReplacementHeuristics: List<VersionReplacementHeuristic>
 }
