+++ kinds/bazel/rules/src/main/kotlin/org/virtuslab/bazelsteward/bazel/rules/BazelRuleHeuristic.kt
@@ -1,21 +1,22 @@
-package org.virtuslab.bazelsteward.core.replacement
+package org.virtuslab.bazelsteward.bazel.rules
 
 import org.virtuslab.bazelsteward.core.common.FileChange
 import org.virtuslab.bazelsteward.core.common.TextFile
 import org.virtuslab.bazelsteward.core.common.UpdateSuggestion
-import org.virtuslab.bazelsteward.core.rules.RuleLibrary
-import org.virtuslab.bazelsteward.core.rules.RuleVersion
+import org.virtuslab.bazelsteward.core.replacement.LibraryUpdate
+import org.virtuslab.bazelsteward.core.replacement.VersionReplacementHeuristic
 
-object BazelRuleHeuristic : Heuristic {
+object BazelRuleHeuristic : VersionReplacementHeuristic {
   override val name: String = "bazel-rule-default"
 
   override fun apply(files: List<TextFile>, updateSuggestion: UpdateSuggestion): LibraryUpdate? {
     if (updateSuggestion.currentLibrary is RuleLibrary && updateSuggestion.suggestedVersion is RuleVersion) {
-      val currentUrl = updateSuggestion.currentLibrary.id.downloadUrl
+      val ruleLibrary = updateSuggestion.currentLibrary as RuleLibrary
+      val currentUrl = ruleLibrary.id.downloadUrl
       val currentVersion = updateSuggestion.currentLibrary.version.value
-      val currentSha = updateSuggestion.currentLibrary.id.sha256
+      val currentSha = ruleLibrary.id.sha256
 
-      val changes = with(updateSuggestion.suggestedVersion) {
+      val changes = with(updateSuggestion.suggestedVersion as RuleVersion) {
         listOf(currentUrl, currentVersion, currentSha).zip(listOf(url, value, sha256)).flatMap { (current, suggested) ->
           val regex = """(${Regex.escape(current)})""".toRegex()
           files
