+++ kinds/bazel/rules/src/test/kotlin/org/virtuslab/bazelsteward/bazel/rules/BazelRuleLibraryIdTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import org.virtuslab.bazelsteward.core.rules.RuleLibraryId
+import org.virtuslab.bazelsteward.bazel.rules.RuleLibraryId
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class BazelRuleLibraryIdTest {
